+++ packages/util/src/index.ts
@@ -7,3 +7,4 @@ export * from './string';
 export * from './type';
 export * from './web';
 export * from './optional';
+export * from './operation';